[#39] Fix Field Naming Conventions for Field Resize
Drupal's convention is to name fields `field_X_value` in a _table_ called `field_data_field_X`; the code was previously assuming that the value field was `field_data_field_X_value`, which is not correct. The name of the field is shorter than the table name (presumably to make more characters available for longer field names on MySQL).